Move plot configs to a separate hook
